refactor: changing component paths
