+++ src/app/globals.css
@@ -5,6 +5,7 @@
 :root {
   background: radial-gradient(40% 130% at 45% 75%, #3A3A3A 0%, #0B0B0B 100%);
   background-repeat: repeat-y;
+  color: white;
 }
 
 @font-face {
