+++ src/app/layout.js
@@ -1,7 +1,7 @@
 import { Inter } from "next/font/google";
-import NavBar from "@/components/layout/NavBar";
+import NavBar from "@/components/layout/Navbar/NavBar";
 import "./globals.css";
-import Footer from "@/components/layout/Footer";
+import Footer from "@/components/layout/Footer/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
